etl/requirements.txt
backoff==1.11.1
certifi==2021.10.8
elastic-transport==8.0.1
elasticsearch==8.0.0
psycopg2-binary==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
python-environ==0.4.54
typing_extensions==4.1.1
urllib3==1.26.8
